Implement most of the stats Stats 3 had.

+++ src/main/java/nl/lolmewn/stats/player/SimpleStatContainer.java
@@ -12,7 +12,7 @@ public class SimpleStatContainer {
 
     private final Disposable subscription;
     private final StatsContainer parent;
-    private final Map<Map<String, Object>, Long> values = new HashMap<>();
+    private final Map<Map<String, Object>, Double> values = new HashMap<>();
 
     public SimpleStatContainer(StatsContainer parent) {
         this.parent = parent;
@@ -23,14 +23,14 @@ public class SimpleStatContainer {
         Map<String, Object> collect = this.parent.getStat().getMetaData().stream().filter(StatMetaData::isGroupable)
                 .map(StatMetaData::getId)
                 .collect(Collectors.toMap(Function.identity(), o -> entry.getMetadata().get(o)));
-        this.values.merge(collect, entry.getAmount(), Long::sum);
+        this.values.merge(collect, entry.getAmount(), Double::sum);
     }
 
-    public Map<Map<String, Object>, Long> getValues() {
+    public Map<Map<String, Object>, Double> getValues() {
         return values;
     }
 
-    public Map<Map<String, Object>, Long> getValuesWhere(String key, Object value) {
+    public Map<Map<String, Object>, Double> getValuesWhere(String key, Object value) {
         return this.getValues().entrySet().stream()
                 .filter(entry -> entry.getKey().containsKey(key) && entry.getKey().get(key).equals(value))
                 .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
